Edited chrome launch command to be universal.
The old launch command for chrome mapped to Program Files (x86), however this is not necessarily where chrome is installed on all machines. The launch command was edited to use progra~2 to auto-map to the system's default installation location. This will prevent the software from not working on some 64-bit machines.